requirements.txt
Django==5.0.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.5.1
gunicorn==22.0.0
whitenoise==6.7.0
